[SUBMITTED 20230811] [RFC] extrawarn: enable more W=1 warnings by default
A number of warning options from the W=1 set are completely clean in current
kernels, so we should just enable them by default, including a lot of warnings
that are part of -Wextra, so just turn on -Wextra by default.

The -Woverride-init, -Wvoid-pointer-to-enum-cast and
-Wmissing-format-attribute warnings are part of -Wextra but still produce
some legitimate warnings that need to be fixed, so leave them at the
W=1 level but turn them off otherwise.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>