[SUBMITTED 20230802] sh: remove superhyway bus support
superhyway was only referenced on sh4-202, which is now gone, so remove
it all as well.

I could find no trace of anything ever calling
superhyway_register_driver(), not in the git history but also not on the
web, so I assume this has never served any purpose on mainline kernels.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>