[SUBMITTED 20230804] watchdog/hardlockup: simplify Kconfig selection
The use of the 'imply' keyword tends to be a bad idea, as it's at best
confusing to the reader but often actually doesn't do what the author
intended.

In this case, it seems to be used correctly, but doing the same thing
using 'default' statements as we have elsewhere in the kernel is simpler
and would be easier to understand by readers that are unfamiliar with
the special semantics of 'imply'.

Fixes: 1356d0b966e7e ("watchdog/hardlockup: make the config checks more straightforward")
Signed-off-by: Arnd Bergmann <arnd@arndb.de>
---
I got confused by the 'imply' here myself and thought this had caused
a build failure that turned out to be unrelated, but it might help
to apply this anyway to save the next person the confusion.