[SUBMITTED 20230814] ASoC: SOF: Intel: fix hda_sdw_check_wakeen_irq()
This function cannot work, as the 'chip' variable is never initialized:

sound/soc/sof/intel/hda.c:423:6: error: variable 'chip' is uninitialized when used here [-Werror,-Wuninitialized]
        if (chip && chip->check_sdw_wakeen_irq)
            ^~~~

Set it the same way that other functions in this file do.

Fixes: 9362ab78f175d ("ASoC: SOF: Intel: add abstraction for SoundWire wake-ups")
Signed-off-by: Arnd Bergmann <arnd@arndb.de>
---
There hasn't been a new linux-next in a few days, so there is a good chance
someone else already fixed this in the meantime. Sending out my fix in case
that hasn't happened yet.