fbdev: move EXPORT_SYMBOL statements next to functions
As a preparation for additional cleanups, make sure each exported
function only has one global definition and the export is right
next to it.

This follows the regular kernel coding style, and allows adding
more #ifdef checks to hide unused functions in configurations that
disable the deprecated CONFIG_FB_DEVICE and/or CONFIG_FB and
only use FB_CORE in order to provide a framebuffer console for
DRM drivers.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>