serial: amba-pl011: stop using phys_to_page()
A patch series from Gregory Clement to use amba drivers on mips fails
to build because of the lack of the phys_to_page() function:

       drivers/tty/serial/amba-pl011.c: In function 'pl011_sgbuf_init':

            drivers/tty/serial/amba-pl011.c:380:30: error: implicit declaration of function 'phys_to_page'; did you mean 'pfn_to_page'? [-Werror=implicit-function-declaration]

         380 |         sg_set_page(&sg->sg, phys_to_page(dma_addr),
             |                              ^~~~~~~~~~~~
             |                              pfn_to_page

I discussed this with Gregory on IRC, and prototyped a
possible fix. The issue was caused by the use of coherent memory
for the buffer and passing that into a scatterlist structure.

Since there is no guarantee that the memory returned by
dma_alloc_coherent() is associated with a 'struct page', using
the architecture specific phys_to_page() is wrong, but using
virt_to_page() would be as well.

An easy workaround is to stop using sg lists altogether and
just use the *_single() functions instead. This also simplifies
the code a bit since the scatterlists in this driver always have
only one entry anyway.

Fixes: cb06ff102e2d7 ("ARM: PL011: Add support for Rx DMA buffer polling.")
Reported-by: kernel test robot <lkp@intel.com>
Closes: https://lore.kernel.org/oe-kbuild-all/202310050726.GDpZbMDO-lkp@intel.com/
Signed-off-by: Arnd Bergmann <arnd@arndb.de>