[RFC] wireless: move obsolete drivers to staging
While looking at the old drivers using the obsolete .ndo_do_ioctl()
callback, I found a number of network drivers that are especially
obsolete, in particular for 802.11b (11Mbit/s) or even older wireless
networks, using non-busmaster ISA/PCMCIA style bus interfaces, and using
the legacy wireless extension ioctls rather than the netlink interfaces
that were meant to replace them in 2007. All of these drivers are
obsolete or orphaned.

We had previously discussed this topic, but nobody ever moved the
files, so I now went through the list to my best knowledge. These
are the drivers that I would classify as "probably unused" by now:

 - Atmel at76c502/at76c504/at76c506 is a PIO-only (PCMCIA, mini-PCI
   and Cardbus) 802.11b driver with incomplete CFG80211 support.
   The related at76c50x USB driver uses MAC80211 and remains.

 - Cisco Aironet is an 802.11b PCMCIA and mini-PCI with limited support
   for Cardbus DMA and for CFG80211.

 - HostAP is an ISA/PCMCIA style 802.11b driver supporting only
   wireless extensions, and some custom ioctls (already removed).
   Some devices include a legacy PCI bridge but no DMA.

 - Aviator/Raytheon is an early PCMCIA driver, apparently predating
   802.11b and only supporting wireless extensions.

 - Planet WL3501 is another PCMCIA driver for pre-802.11b interfaces
   (2Mbit/s) with incomplete CFG80211 support

 - Zydas zd1201 is a USB 802.11b driver with limited support for
   CFG80211.

 - Orinoco is a PIO-only ISA/PCMCIA 802.11b device with extra bus
   interface connections for PCI/Cardbus/mini-PCI and a few
   pre-2002 Apple PowerMac variants. It supports both
   wireless extensions and CFG80211, but I could not tell if
   it requires using both.

 - Wireless RNDIS USB is a new-style CFG80211 driver for 802.11b
   and 802.11g USB hardware from around 2004 to 2006. This makes it
   more modern than any of the others, but Kalle already classified
   it as "legacy" in commit 298e50ad8eb8f ("wifi: move raycs, wl3501
   and rndis_wlan to legacy directory"), so it stays with ray_cs and
   wl3501_cs.

There are a few other drivers that are similar to these but that are
more likely to still be needed, and are not moved here:

 - Intel ipw2x00 is a PCI bus-master device for 802.11a/b/g that was
   popular in "Centrino" branded laptops from 2003 to 2005, but it
   still requires wireless extensions.

 - Marvell Libertas is an 802.11a/b/g device with a number of bus
   interfaces (USB, SDIO, SPI, PCMCIA) and incomplete CFG80211
   support. This one was used in the OLPC XO laptop and some other
   embedded devices that are still supported.

 - Some broadcom b43xx devices use the SSB bus that can be abstracted
   through PCMCIA. All of them use CFG80211.

 - The Sony Playstation 3 "gelic" ethernet driver contains a bridge
   and an 802.11b/g client chip that is controlled through a hypervisor
   interface from the OS, and it uses wireless extensions in the kernel
   driver.

Link: https://lore.kernel.org/all/87imaeg4ar.fsf@codeaurora.org/
Signed-off-by: Arnd Bergmann <arnd@arndb.de>
---
I'm mainly taking educated guesses about which drivers are actually
obsolete. Let me know if I got any of them wrong, or if any others
should be included here.

I did not move the staging drivers for ks7010, rtl8192e, rtl8192u,
rtl8712, rtl8723bs, vt6656 or wlan-ng into driver/staging/wireless,
as they are already directly under drivers/staging, but that could
be another patch if we want to keep them all in one place.