[RFC] wifi: remove ipw2100/ipw2200 drivers
These two drivers were used for the earliest "Centrino" branded Intel
laptops during the late 32-bit Pentium-M era, roughly 2003 to 2005, which
probably makes it the most modern platform that still uses the wireless
extension interface instead of cfg80211. Unlike the other drivers that
are suggested for removal, this one is still officially maintained.

According to Johannes Berg, there was an effort to finish the move away
from wext in the past, but the last evidence of this that I could find
is from commit a3caa99e6c68f ("libipw: initiate cfg80211 API conversion
(v2)") in 2009.

Link: https://lore.kernel.org/all/87fs2fgals.fsf@kernel.org/
Cc: Stanislav Yakovlev <stas.yakovlev@gmail.com>
Cc: Linux Wireless <ilw@linux.intel.com>
Signed-off-by: Arnd Bergmann <arnd@arndb.de>
---
I'm not convinced this should be in the same set of drivers as the
rest, since this is clearly less obsolete than the other hardware
that I would remove support for.