[v2] media: pci: mgb4: remove bogus 'select' statements
As this is just a regular device driver, it has no business force-enabling
other drivers in the system, it should be entirely independent of the
implementation of the spi-nor layer or the specific DMA engine.

The IIO symbols that are selected here are library modules that are
legitimately used.

Since mgb4 links against the xdma driver, add a dependency for that
instead of the 'select'.

Fixes: 0ab13674a9bd ("media: pci: mgb4: Added Digiteq Automotive MGB4 driver")
Signed-off-by: Arnd Bergmann <arnd@arndb.de>