mmc: simplify LEDS_CLASS dependencies
There are three mmc drivers that interact with the LED subsystem,
but they all do this differently:

 - au1xmmc.c only uses LED_CLASS if that is built-in, but ignores it
   if it's built as a module.

 - sdhci uses the IS_REACHABLE() helper that lets it use the LED
   also when both LEDS_CLASS and sdhci are loadable modules, but
   may still surprise users by not using the LEDs if SDHCI is
   built-in and LEDS_CLASS is a module.

 - rtsx has the same behavior SDHCI but open-codes it instead of
   using the macro.

Change all three of these to instead depend on the new LEDS_CLASS_OPTIONAL
symbol and use LEDS_CLASS whenever it is enabled. As a result, it the
three drivers can no longer be built-in when LEDS_CLASS=m, avoiding
any ambiguity for users.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>