ARM: deprecate ARCH_DOVE
The Marvell Armada 510 (Dove) platform is supported both as devicetree
and atags/boardfile variants, with varying degrees of hardware support.

Russell was the last known user of the board file version, and there
are very few users overall. Unless Russell still needs this variant,
let's remove it soon, in order to simplify the DT conversion of the
remaining plat-orion machines (mv78xx0 and orion5x).

Signed-off-by: Arnd Bergmann <arnd@arndb.de>