pc104: tighten Kconfig dependencies
CONFIG_PC104 can be enabled on any architecture at the moment,
but most do not actually have the logic for hardcoded ISA I/O
ports.

Add a Kconfig dependency to PC104 to allow it only on x86 as well
as a couple of specific machines that are known to have PC104
connectors. This list can be extended if there are others of course,
but it makes sense to have this as an explicit opt-in.

In particular, only configurations that provide CONFIG_HAS_IOPORT
should ever be able to use PC104 add-on boards.

In drivers/gpio, there are two chips that select CONFIG_ISA_BUS_API
without a dependency. Like the other users of ISA_BUS_API, these
seem to be tied PC104, so add another dependency there.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>