tty: serial: remove CONFIG_SERIAL_8250_RUNTIME_UARTS
The original purpose of this option was to allow the kernel to support
a larger compile-time CONFIG_SERIAL_8250_NR_UARTS without all of the
devices actually getting registered at boot time, but instead set
it as a boot time "nr_uarts=" option.

In Linux-6.5, another change was done to do the same thing better,
allowing up to CONFIG_SERIAL_8250_NR_UARTS uarts to be registered
regardless of the boot time option, but at this point there was
no longer a need to set CONFIG_SERIAL_8250_RUNTIME_UARTS either.

Change the default nr_uarts to be the same as the set of initial
ISA ports on x86/alpha/m68k, or otherwise zero, initializing only
the ports that are defined on a given platform.

Fixes: a61c2d78ce61 ("[SERIAL] Make the number of UARTs registered configurable.")
Fixes: 9d86719f8769 ("serial: 8250: Allow using ports higher than SERIAL_8250_RUNTIME_UARTS")
Signed-off-by: Arnd Bergmann <arnd@arndb.de>